Added unique message IDs per-client and message acknowledgement.
Message IDs:
- Client to server message IDs start with 0 and increment by 2 each message (so are only even message IDs)
- Server to client message IDs start with 1 and increment by 2 each message (so are only odd message IDs)

Message acknowledgements:
- Each message (exluding disconnect) is responded with the same message ID but with message type 1.
- Messages that were sent and acknowledged, sent but not acknowledged, failed to send, and were received are stored in each client as List<Message> messageSentAcknowledged, List<Message> messageSentNotAcknowledged, List<Message> messageToBeSent, and List<Message> messageReceived respectively.

Handling of messages that failed to send or were not acknowledged has yet to be implemented; these messages should be resent periodically until acknowledged, using the same message ID as the original attempt.

Messages that have already been received and processed, with an incoming message of the same message ID, should be acknowledged and then ignored.