Dispose of certificate after use
Not disposing of a certificate file loaded directly from a file would result in an encrypted version of the certificate being stored in %ALLUSERSPROFILE%\Application Data\Microsoft\Crypto\RSA\S-1-5-18 (can change depending on user account, context...).

Dispose of the certificate once it is finished being used. This will not be called if the program is forcibly closed.