Replace references of TCPClient to SLStream
Closing only the TCPClient won't close the SSlStream, which might cause issues. Checking if the TCPClient is null won't check if the SSLStream is null, which is more important.

Close the SSLStream instead, which will automatically close the underlying TCPClient as well. Check if the SSLStream is null instead of the TCPClient.