Store permission number as a bit-flag Lots of memory is needed to store a collection of individual permission numbers in order to determine what permissions a user has. As each number is double that of the previous, double the number of bits is needed for every individual permission that must be stored.
It is more efficient to instead use every single bit to store a combination of all possible values, using the FlagsAttribute attribute, [Flags]. Bitwise operations (OR, AND, XOR; represented as |, & and ^) can then be used to determine all the permissions that a single permission number has.

Store permission number as a bit-flag.